+++ tars/tasks/html/minify-html.js
@@ -2,35 +2,43 @@
 
 var gulp = tars.packages.gulp;
 var gutil = tars.packages.gutil;
-var htmlMin = tars.packages.htmlMin;
-var notify = tars.packages.notify;
+var gulpif = tars.packages.gulpif;
+var minify = tars.packages.htmlMin;
+var prettify = tars.packages.htmlPrettify;
+var plumber = tars.packages.plumber;
 var notifier = tars.helpers.notifier;
 
-var opts = {
+var minifyOpts = {
     conditionals: true,
     quotes: true,
     empty: true
+};
+
+var prettifyOpts = {
+    indent_char: ' ',
+    indent_size: 4,
+    indent_inner_html: true,
+    preserve_newlines: true,
+    max_preserve_newlines: 0,
+    unformatted: true,
+    end_with_newline: true
 };
 
 /**
  * Minify HTML (optional task)
  */
 module.exports = function () {
-    return gulp.task('html:minify-html', function (cb) {
-
-        if (tars.config.minifyHtml) {
-            return gulp.src('./dev/**/*.html')
-                .pipe(htmlMin(opts))
-                .on('error', notify.onError(function (error) {
-                    return '\nAn error occurred while minifing html-files.\nLook in the console for details.\n' + error;
+    return gulp.task('html:minify-html', function () {
+        return gulp.src('./dev/**/*.html')
+                .pipe(plumber({
+                    errorHandler: function (error) {
+                        notifier.error('An error occurred while processing compiled html-files.', error);
+                    }
                 }))
+                .pipe(gulpif(tars.config.minifyHtml, minify(minifyOpts), prettify(prettifyOpts)))
                 .pipe(gulp.dest('./dev/'))
                 .pipe(
-                    notifier('Html \'ve been minified')
+                    notifier.success('Compiled html\'ve been processed.')
                 );
-        } else {
-            gutil.log('!Html-minify disabled!');
-            cb(null);
-        }
     });
 };