+++ tars/tasks/images/minify-raster-img.js
@@ -4,24 +4,28 @@ var gulp = tars.packages.gulp;
 var gutil = tars.packages.gutil;
 var imagemin = tars.packages.imagemin;
 var changed = tars.packages.changed;
-var notify = tars.packages.notify;
+var plumber = tars.packages.plumber;
 var notifier = tars.helpers.notifier;
+
+var staticFolderName = tars.config.fs.staticFolderName;
+var imagesFolderName = tars.config.fs.imagesFolderName;
 
 /**
  * Minify png and jpg images
  */
 module.exports = function () {
     return gulp.task('images:minify-raster-img', function (cb) {
-        return gulp.src('./dev/' + tars.config.fs.staticFolderName + '/' + tars.config.fs.imagesFolderName + '/**/*.{png, jpg}')
-            .pipe(changed('./dev/' + tars.config.fs.staticFolderName + '/' + tars.config.fs.imagesFolderName + '/'))
+        return gulp.src('./dev/' + staticFolderName + '/' + imagesFolderName + '/**/*.{png, jpg}')
+            .pipe(plumber({
+                errorHandler: function (error) {
+                    notifier.error('An error occurred while minifying raster images.', error);
+                }
+            }))
+            .pipe(changed('./dev/' + staticFolderName + '/' + imagesFolderName + '/'))
             .pipe(imagemin())
-            .on('error', notify.onError(function (error) {
-                    return '\nAn error occurred while minifying raster images.\nLook in the console for details.\n' + error;
-                })
-            )
-            .pipe(gulp.dest('./dev/' + tars.config.fs.staticFolderName + '/' + tars.config.fs.imagesFolderName + '/'))
+            .pipe(gulp.dest('./dev/' + staticFolderName + '/' + imagesFolderName + '/'))
             .pipe(
-                notifier('Rastered images\'ve been minified')
+                notifier.success('Rastered images\'ve been minified')
             );
     });
 };