+++ tars/tasks/images/minify-svg.js
@@ -4,8 +4,11 @@ var gulp = tars.packages.gulp;
 var gutil = tars.packages.gutil;
 var imagemin = tars.packages.imagemin;
 var changed = tars.packages.changed;
-var notify = tars.packages.notify;
+var plumber = tars.packages.plumber;
 var notifier = tars.helpers.notifier;
+
+var staticFolderName = tars.config.fs.staticFolderName;
+var imagesFolderName = tars.config.fs.imagesFolderName;
 
 /**
  * Minify svg-images (optional task)
@@ -13,9 +16,14 @@ var notifier = tars.helpers.notifier;
 module.exports = function () {
     return gulp.task('images:minify-svg', function (cb) {
         if (tars.config.useSVG) {
-            return gulp.src('./markup/' + tars.config.fs.staticFolderName + '/' + tars.config.fs.imagesFolderName + '/svg/*.svg')
+            return gulp.src('./markup/' + staticFolderName + '/' + imagesFolderName + '/svg/*.svg')
+                .pipe(plumber({
+                    errorHandler: function (error) {
+                        notifier.error('An error occurred while minifying svg.', error);
+                    }
+                }))
                 .pipe(changed(
-                        'dev/' + tars.config.fs.staticFolderName + '/' + tars.config.fs.imagesFolderName + '/minified-svg',
+                        'dev/' + staticFolderName + '/' + imagesFolderName + '/minified-svg',
                         {
                             hasChanged: changed.compareLastModifiedTime,
                             extension: '.svg'
@@ -34,12 +42,9 @@ module.exports = function () {
                         }
                     )
                 )
-                .on('error', notify.onError(function (error) {
-                    return '\nAn error occurred while minifying svg.\nLook in the console for details.\n' + error;
-                }))
-                .pipe(gulp.dest('./dev/' + tars.config.fs.staticFolderName + '/' + tars.config.fs.imagesFolderName + '/minified-svg/'))
+                .pipe(gulp.dest('./dev/' + staticFolderName + '/' + imagesFolderName + '/minified-svg/'))
                 .pipe(
-                    notifier('SVG\'ve been minified')
+                    notifier.success('SVG\'ve been minified')
                 );
         } else {
             gutil.log('!SVG is not used!');