+++ tars/tasks/images/raster-svg.js
@@ -5,9 +5,12 @@ var gutil = tars.packages.gutil;
 var cache = tars.packages.cache;
 var changed = tars.packages.changed;
 var svg2png = tars.packages.svg2png;
-var notify = tars.packages.notify;
+var plumber = tars.packages.plumber;
 var notifier = tars.helpers.notifier;
 var browserSync = tars.packages.browserSync;
+
+var staticFolderName = tars.config.fs.staticFolderName;
+var imagesFolderName = tars.config.fs.imagesFolderName;
 
 /**
  * Raster SVG-files (optional task)
@@ -15,12 +18,17 @@ var browserSync = tars.packages.browserSync;
 module.exports = function () {
     return gulp.task('images:raster-svg', function (cb) {
 
-        if (tars.config.useSVG && tars.flags.ie8) {
-            return gulp.src('./markup/' + tars.config.fs.staticFolderName + '/' + tars.config.fs.imagesFolderName + '/svg/*.svg')
+        if (tars.config.useSVG && (tars.flags.ie8 || tars.flags.ie)) {
+            return gulp.src('./markup/' + staticFolderName + '/' + imagesFolderName + '/svg/*.svg')
+                .pipe(plumber({
+                    errorHandler: function (error) {
+                        notifier.error('An error occurred while rastering svg.', error);
+                    }
+                }))
                 .pipe(cache('raster-svg'))
                 .pipe(
                     changed(
-                        'dev/' + tars.config.fs.staticFolderName + '/' + tars.config.fs.imagesFolderName + '/rastered-svg-images',
+                        'dev/' + staticFolderName + '/' + imagesFolderName + '/rastered-svg-images',
                         {
                             hasChanged: changed.compareLastModifiedTime,
                             extension: '.png'
@@ -28,12 +36,9 @@ module.exports = function () {
                     )
                 )
                 .pipe(svg2png())
-                .on('error', notify.onError(function (error) {
-                    return '\nAn error occurred while rastering svg.\nLook in the console for details.\n' + error;
-                }))
-                .pipe(gulp.dest('dev/' + tars.config.fs.staticFolderName + '/' + tars.config.fs.imagesFolderName + '/rastered-svg-images'))
+                .pipe(gulp.dest('dev/' + staticFolderName + '/' + imagesFolderName + '/rastered-svg-images'))
                 .pipe(
-                    notifier('SVG\'ve been rastered')
+                    notifier.success('SVG\'ve been rastered')
                 );
         } else {
             gutil.log('!Rastering SVG is not used!');