+++ tars/tasks/other/move-misc-files.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var gulp = tars.packages.gulp;
-var notify = tars.packages.notify;
+var plumber = tars.packages.plumber;
 var notifier = tars.helpers.notifier;
 var browserSync = tars.packages.browserSync;
 
@@ -11,13 +11,15 @@ var browserSync = tars.packages.browserSync;
 module.exports = function () {
     return gulp.task('other:move-misc-files', function (cb) {
         return gulp.src('./markup/' + tars.config.fs.staticFolderName + '/misc/**/*.*')
-            .on('error', notify.onError(function (error) {
-                return '\nAn error occurred while moving misc-files.\nLook in the console for details.\n' + error;
+            .pipe(plumber({
+                errorHandler: function (error) {
+                    notifier.error('An error occurred while moving misc-files.', error);
+                }
             }))
             .pipe(gulp.dest('./dev/'))
             .pipe(browserSync.reload({ stream: true }))
             .pipe(
-                notifier('Misc files\'ve been moved')
+                notifier.success('Misc files\'ve been moved')
             );
     });
 };