+++ tars/user-tasks/example-task.js
@@ -4,12 +4,13 @@
 
 var gulp = tars.packges.gulp;
 var tarsConfig = tars.config;
-var notify = tars.packages.notify;
+var plumber = tars.packages.plumber;
 var notifier = tars.helpers.notifier;
-// Include browserSync, if you need to reload browser
+// Include browserSync, if you need to reload browser:
 // var browserSync = tars.packages.browserSync;
 
-// require('./ path to task file, which have to be done before current task');
+// require('path to task file, which have to be done before current task');
+// For example:
 // require('./required-task-name');
 
 /**
@@ -19,19 +20,20 @@ module.exports = function () {
 
     return gulp.task('task-name', /*['required-task-name'],*/ function (cb) {
         return gulp.src(/* path-string or array of path-strings to files */)
-            // Do stuff here
-            .on('error', notify.onError(function (error) {
-                return '\nAn error occurred while something.\nLook in the console for details.\n' + error;
+            .pipe(plumber({
+                errorHandler: function (error) {
+                    notifier.error('An error occurred while something.', error);
+                }
             }))
+            // Do stuff here, like
+            // .pipe(less())
             .pipe(gulp.dest(/* path-string to destination directory. Only directory, not a file! */))
 
             // If you need to reload browser, uncomment the row below
             // .pipe(browserSync.reload({stream:true}))
             .pipe(
                 // You can change text of success message
-                notifier('Example task is finished')
-                // if you need notify after each file will be processed, you have to use
-                // notifier('Example task is finished', false)
+                notifier.success('Example task has been finished')
             );
 
         // You can return callback, if you can't return pipe